% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_autoplot.R
\name{autoplot.evalcurve}
\alias{autoplot.evalcurve}
\title{Plot the result of Precision-Recall curve validation}
\usage{
\method{autoplot}{evalcurve}(object, base_plot = TRUE, ret_grob = FALSE,
  ...)
}
\arguments{
\item{object}{An S3 object that contains evaluation results of
Precision-Recall curves.}

\item{base_plot}{A Boolean value to specify whether the base points are
plotted.}

\item{ret_grob}{A Boolean value to specify whether the function returs a
grob object.}

\item{...}{Not used by this function.}
}
\value{
A data frame with validation results.
}
\description{
The \code{plot_eval_results} function validate Precision-Recall curves
   and creat a plot.
}
\examples{
library(ggplot2)

## Plot evaluation results on test datasets r1, r2, and r3
eres1 <- eval_curves(c("r1", "r2", "r3"), "crv")
autoplot(eres1)

}

