% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_sets}
\alias{benchmark_sets}
\title{Run microbenchmark with specified sets of tools and sets of sample datasets}
\usage{
benchmark_sets(sampset_names = c("b100", "ib100", "b10k", "ib10k"),
  toolset_names = c("crv", "auc", "both"), times = 5, unit = "ms")
}
\arguments{
\item{sampset_names}{A character vector to specify the names of sample sets
for \code{\link{create_sample}}.}

\item{toolset_names}{A character vector to specify the names of tool sets for
\code{\link{create_tools}}.}

\item{times}{The number of iteration used in
\code{\link[microbenchmark]{microbenchmark}}.}

\item{unit}{A single string to specify the unit used in
\code{\link[microbenchmark]{summary.microbenchmark}}.}
}
\value{
A data frame with microbenchmark results.
}
\description{
The \code{benchmark_sets} function runs
    \code{\link[microbenchmark]{microbenchmark}} for specified sets of tools
    and sets of sample datasets
}
\examples{
## Generate a sample dataset
res1 <- benchmark_sets(c("b100", "ib100"), c("crv", "auc", "both"))

}
\seealso{
\code{\link{run_benchmark}} is used for benchmarking.
   \code{\link[microbenchmark]{microbenchmark}} for benchmarking details.
   \code{\link{create_sample}} to generate sample datasets.
   \code{\link{create_tools}} to create sets of tools.
}

