% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{data}
\name{C1DATA}
\alias{C1DATA}
\title{C1: Pre-calculated Precision-Recall curve}
\format{A list with 5 items.
\describe{
  \item{scores}{input scores}
  \item{labels}{input labels}
  \item{bp_x}{pre-calculated recall values for curve evaluation}
  \item{bp_y}{pre-calculated precision values for curve evaluation}
  \item{tp_x}{x position for displaying the test result in a plot}
  \item{tp_y}{y position for displaying the test result in a plot}
}}
\usage{
data(C1DATA)
}
\description{
A list contains scores, labels, and pre-calculated recall and precision
values as x and y.
}
\keyword{datasets}

