% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{data}
\name{C3DATA}
\alias{C3DATA}
\title{C2: Pre-calculated Precision-Recall curve}
\format{See \code{\link{C1DATA}}.}
\usage{
data(C3DATA)
}
\description{
A list contains scores, labels, and precalculated recall and precision
values as x and y.
}
\keyword{datasets}

