% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{data}
\name{C4DATA}
\alias{C4DATA}
\title{C4: Pre-calculated Precision-Recall curve}
\format{See \code{\link{C1DATA}}.}
\usage{
data(C4DATA)
}
\description{
A list contains scores, labels, and pre-calculated recall and precision
values as x and y.
}
\keyword{datasets}

