% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_interface.R
\name{create_example_func}
\alias{create_example_func}
\title{Create an example for the func argument of the create_usrtool function}
\usage{
create_example_func()
}
\value{
A function as an example for \code{\link{create_usrtool}}
}
\description{
The \code{create_example_func} function creates an example for the
 \code{\link{create_usrtool}} function.
}
\examples{
## Create a function
func <- create_example_func()
func

}
\seealso{
\code{\link{create_usrtool}} requires the same format.
  \code{\link{create_testset}} for \code{testset}.
}

