% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{create_rnd_sample}
\alias{create_rnd_sample}
\title{Create random samples}
\usage{
create_rnd_sample(np = 10, pfunc = NULL, nn = 10, nfunc = NULL)
}
\arguments{
\item{np}{An integer to specify the number of positves.}

\item{pfunc}{A function to generate randome positive scores. It should accept
\code{n} as an argument for the number of observations.}

\item{nn}{An integer to specify the number of negatives.}

\item{nfunc}{A function to generate randome negative scores. It should accept
\code{n} as an argument for the number of observations.}
}
\value{
PRCData \code{R6} object.
}
\description{
The \code{create_rnd_sample} function creates a random sample.
}
\examples{
## Generate a sample dataset
samp1 <- create_rnd_sample()

## Generate a sample dataset with 1000 positives and 10000 negatives
samp2 <- create_rnd_sample(np = 1000, nn = 10000)

## Generate a sample dataset by normal distribution random generator
pfunc <- function(n) {rnorm(n)}
samp3 <- create_rnd_sample(pfunc = pfunc)

}
\seealso{
\code{\link{PRCData}} and \code{\link{create_sampleset}}.
}

