% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{create_rnd_samples}
\alias{create_rnd_samples}
\title{Create random samples}
\usage{
create_rnd_samples(np = 10, pfunc = NULL, nn = 10, nfunc = NULL)
}
\arguments{
\item{np}{An integer to specify the number of positves.}

\item{pfunc}{A function to generate randome positive scores.}

\item{nn}{An integer to specify the number of negatives.}

\item{pfunc}{A function to generate randome negative scores.}
}
\value{
PRCData \code{R6} object.
}
\description{
The \code{create_rnd_samples} function creates a random sample.
}
\examples{
## Generate a sample dataset
samp1 <- create_rnd_samples()

## Generate a sample dataset with 1000 positives and 10000 negatives
samp2 <- create_rnd_samples(np = 1000, nn = 10000)

}
\seealso{
\code{\link{PRCData}} is internally generated by
  \code{\link{prcdata_generator}}.
}

