% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{create_sampleset}
\alias{create_sampleset}
\title{Create a set of random samples}
\usage{
create_sampleset(set_name)
}
\arguments{
\item{set_name}{A single string to specify the name of a predefined tool set.
  Following six sets are currently available.

  \describe{
    \item{"b100"}{A balanced data set with 50 positives and 50
        negatives.}
    \item{"b10k"}{A balanced data set with 5000 positives and 5000
        negatives.}
    \item{"b100k"}{A balanced data set with 50 000 positives and 50 000
        negatives}
    \item{"ib100"}{An imbalanced data set with 25 positives and 75
        negatives.}
    \item{"ib10k"}{A imbalanced data set with 2500 positives and 7500
        negatives.}
    \item{"ib100k"}{A imbalanced data set with 25 000 positives and 75 000
        negatives}
  }}
}
\value{
\code{\link{PRCData}} is generated by
   \code{\link{prcdata_generator}}.
}
\description{
The \code{create_sampleset} function takes a name of predefined sample set
  and genarates a list of wrapper functions for tool calculations.
}
\examples{
## Create a balanced data set with 50 positives and 50 negatives
samp1 <- create_sampleset("b100")

## Create a balanced data set with 50 0000 positives and 50 000 negatives
samp2 <- create_sampleset("b100k")

## Create an imbalanced data set with 25 positives and 75 negatives
samp3 <- create_sampleset("ib100")

}
\seealso{
\code{\link{PRCData}} and \code{\link{prcdata_generator}}.
}

