% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{create_testdata}
\alias{create_testdata}
\title{Create a set of random samples}
\usage{
create_testdata(datatype, sampnames = NULL, scores = NULL, labels = NULL,
  dsname = NA, base_x = NULL, base_y = NULL, text_x = NULL,
  text_y = NULL)
}
\arguments{
\item{samp_name}{A single string to specify the name of a predefined tool
  set. Following six sets are currently available.

  \describe{
    \item{"b100"}{A balanced data set with 50 positives and 50
        negatives.}
    \item{"b10k"}{A balanced data set with 5000 positives and 5000
        negatives.}
    \item{"b100k"}{A balanced data set with 50 000 positives and 50 000
        negatives}
    \item{"ib100"}{An imbalanced data set with 25 positives and 75
        negatives.}
    \item{"ib10k"}{A imbalanced data set with 2500 positives and 7500
        negatives.}
    \item{"ib100k"}{A imbalanced data set with 25 000 positives and 75 000
        negatives}
  }}

\item{np}{An integer to specify the number of positves.}

\item{pfunc}{A function to generate randome positive scores. It should accept
\code{n} as an argument for the number of observations.}

\item{nn}{An integer to specify the number of negatives.}

\item{nfunc}{A function to generate randome negative scores. It should accept
\code{n} as an argument for the number of observations.}
}
\description{
The \code{create_sample} function takes a name of predefined sample set
  and genarates a list of wrapper functions for tool calculations.
}
\examples{
## Create a balanced data set with 50 positives and 50 negatives
samp1 <- create_sample("b100")

## Create a balanced data set with 50 0000 positives and 50 000 negatives
samp2 <- create_sample("b100k")

## Create an imbalanced data set with 25 positives and 75 negatives
samp3 <- create_sample("ib100")

## Generate a sample dataset with 1000 positives and 10000 negatives
samp4 <- create_sample(np = 1000, nn = 10000)

## Generate a sample dataset by normal distribution random generator
pfunc <- function(n) {rnorm(n)}
samp5 <- create_sample(pfunc = pfunc)

}
\seealso{
\code{\link{PRCData}}
}

