% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_interface.R
\name{create_tool}
\alias{create_tool}
\title{Create tool wrapper objects}
\usage{
create_tool(tool_names = c("ROCR", "AUCCalculator", "PerfMeas", "PRROC",
  "precrec"), init_params = replicate(length(tool_names), list()))
}
\arguments{
\item{tool_names}{A character vector to specify the names of
  performance evaluatoin tools. The names for the following five tools can be
  currently used.

  \itemize{
    \item ROCR
    \item AUCCalculator
    \item PerfMeas
    \item PRROC
    \item precrec
  }}

\item{init_params}{A list of lists that contain parameters used for
  tool object intialization. For instance, \code{list(list(p1 = 1))} makes
  the first tool (only one tool in this case) is initialized with a
  parameter as \code{toolclass$new(p1 = 1)}.

  \tabular{lll}{
    \strong{Tool}
    \tab \strong{Parameters}
    \tab \strong{Notes} \cr

    ROCR
        \tab NA \tab NA   \cr
    AUCCalculator
        \tab \code{type}, \code{fpath}
        \tab see \code{\link{ToolAUCCalculator}} \cr
    PerfMeas
        \tab NA \tab NA \cr
    PRROC
        \tab \code{curve}, \code{minStepSize}
        \tab see \code{\link{ToolPRROC}} \cr
    precrec
        \tab NA \tab NA
  }}
}
\value{
A list of \code{R6} objects.
}
\description{
The \code{create_tool} function takes names of performance evaluation
  tool and generate wrapper objects.
}
\examples{
## Generate all wraper objects
tools1 <- create_tool()

## Generate PRROC and precrec wraper objects
tools2 <- create_tool(c("PRROC", "precrec"))

## Specifiy an initialize parameter for the second tool entry
tools3 <- create_tool(c("PRROC", "PRROC"),
                      list(list(), list(curve = FALSE)))

}
\seealso{
\code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
  \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}},
  and \code{\link{Toolprecrec}} can be genereted by this function.
}

