% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_interface.R
\name{create_tools}
\alias{create_tools}
\title{Create a set of tools}
\usage{
create_tools(tool_names = NULL, init_params = NULL, retval = TRUE,
  auc = TRUE)
}
\arguments{
\item{tool_names}{A character vector to specify the names of
  performance evaluatoin tools. The names for the following five tools can be
  currently used.

  \itemize{
    \item ROCR
    \item AUCCalculator
    \item PerfMeas
    \item PRROC
    \item precrec
  }

  Alternatively, a pre-defined set name can be specified. Following three
  sets are currently available.

  \describe{
    \item{"crv"}{A set of tools optimized for curve calculations}
    \item{"auc"}{A set of tools optimized for AUC calculations}
    \item{"both"}{A set of tools optimized for both curve and AUC
                  calculations.}
  }}

\item{init_params}{A list of lists that contain parameters used for
  tool object intialization. For instance, \code{list(list(p1 = 1))} makes
  the first tool (only one tool in this case) is initialized with a
  parameter as \code{toolclass$new(p1 = 1)}.

  \tabular{lll}{
    \strong{Tool}
    \tab \strong{Parameters}
    \tab \strong{Notes} \cr

    ROCR
        \tab NA \tab NA   \cr
    AUCCalculator
        \tab \code{type}, \code{fpath}
        \tab see \code{\link{ToolAUCCalculator}} \cr
    PerfMeas
        \tab NA \tab NA \cr
    PRROC
        \tab \code{curve}, \code{minStepSize}
        \tab see \code{\link{ToolPRROC}} \cr
    precrec
        \tab NA \tab NA
  }}

\item{retval}{A Boolean value to specify whether the method returns the AUC
score and x-y positions.}

\item{auc}{A Boolean value to specify  whether the AUC score should be
calculated.}
}
\value{
A list of functions that calls tool calculations.
}
\description{
The \code{create_tools} function takes a name of predefined tool set and
  genarates a list of wrapper functions for tool calculations.
}
\examples{
## Create a tool set for curve calculations
tool_funcs1 <- create_tools("crv")

## Create a tool set for AUC calculations
tool_funcs2 <- create_tools("auc")

## Create a tool set for both curve and AUC calculations
tool_funcs3 <- create_tools("both")

}
\seealso{
\code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
  \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}},
  and \code{\link{Toolprecrec}} as classes of tool objects.
}

