% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_interface.R
\name{create_toolset}
\alias{create_toolset}
\title{Create a set of tools}
\usage{
create_toolset(
  tool_names = NULL,
  set_names = NULL,
  calc_auc = TRUE,
  store_res = TRUE
)
}
\arguments{
\item{tool_names}{A character vector to specify the names of
  performance evaluation tools. The names for the following seven tools can
  be currently used.

  \itemize{
    \item ROCR
    \item AUCCalculator
    \item PerfMeas
    \item PRROC
    \item precrec
    \item yardstick
    \item sklearn
  }

  The \code{sklearn} tool requires the \code{reticulate} package together
  with a working Python installation and \code{numpy}. When they are
  unavailable, the tool is created without them and returns a flat dummy
  curve, in the same way as \code{AUCCalculator} does without \code{rJava}.}

\item{set_names}{A character vector to specify a predefined set name.
  Following twelve sets are currently available. The digit is the number of
  tools in the set, and each smaller set drops one more of the slower tools.

  \describe{
    \item{"def7"}{ROCR, AUCCalculator, PerfMeas, PRROC, precrec, yardstick,
      and sklearn with \code{calc_auc = TRUE} and \code{store_res = TRUE}}
    \item{"auc7"}{The 7 tools of "def7" with \code{calc_auc = TRUE}
      and \code{store_res = FALSE}}
    \item{"crv7"}{The 7 tools of "def7" with \code{calc_auc = FALSE}
      and \code{store_res = TRUE}}
    \item{"def6"}{ROCR, AUCCalculator, PRROC, precrec, yardstick, and
      sklearn with \code{calc_auc = TRUE} and \code{store_res = TRUE}}
    \item{"auc6"}{The 6 tools of "def6" with \code{calc_auc = TRUE}
      and \code{store_res = FALSE}}
    \item{"crv6"}{The 6 tools of "def6" with \code{calc_auc = FALSE}
      and \code{store_res = TRUE}}
    \item{"def5"}{ROCR, PRROC, precrec, yardstick, and sklearn with
      \code{calc_auc = TRUE} and \code{store_res = TRUE}}
    \item{"auc5"}{The 5 tools of "def5" with \code{calc_auc = TRUE}
      and \code{store_res = FALSE}}
    \item{"crv5"}{The 5 tools of "def5" with \code{calc_auc = FALSE}
      and \code{store_res = TRUE}}
    \item{"def4"}{ROCR, precrec, yardstick, and sklearn with
      \code{calc_auc = TRUE} and \code{store_res = TRUE}}
    \item{"auc4"}{The 4 tools of "def4" with \code{calc_auc = TRUE}
      and \code{store_res = FALSE}}
    \item{"crv4"}{The 4 tools of "def4" with \code{calc_auc = FALSE}
      and \code{store_res = TRUE}}
  }}

\item{calc_auc}{A Boolean value to specify whether the AUC score should be
calculated.}

\item{store_res}{A Boolean value to specify whether the calculated curve is
retrieved and stored}
}
\value{
A list of \code{R6} tool objects.
}
\description{
The \code{create_toolset} function takes names of predefined tools and
  generates a list of wrapper functions for Precision-Recall curve
  calculations.
}
\examples{
## Create ROCR and precrec
toolset1 <- create_toolset(c("ROCR", "precrec"))
toolset1

## Create auc7 tools
toolset2 <- create_toolset(set_names = "auc7")
toolset2

}
\seealso{
\code{\link{run_benchmark}} and \code{\link{run_evalcurve}} require
 the list of the tools generated by this function
 \code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
 \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}},
 \code{\link{Toolprecrec}}, \code{\link{Toolyardstick}}, and
 \code{\link{Toolsklearn}} as R6 tool classes.
}
