% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{create_usrdata}
\alias{create_usrdata}
\title{Create a user-defined test dataset}
\usage{
create_usrdata(test_type, scores = NULL, labels = NULL, dsname = NA,
  base_x = NULL, base_y = NULL, text_x = NULL, text_y = NULL)
}
\arguments{
\item{test_type}{A single string to specify the type of dataset generated by
  this function.

  \describe{
    \item{"bench"}{Create a test dataset for benchmarking}
    \item{"curve"}{Create a test dataset for curve evaluation}
  }}

\item{scores}{A numeric vector to set scores.}

\item{labels}{A numeric vector to set labels.}

\item{dsname}{A single string to specify the name of the dataset.}

\item{base_x}{A numeric vector to set precalculated recall values for
curve evaulation.}

\item{base_y}{A numeric vector to set precalculated precision values for
curve evaulation.}

\item{text_x}{A single numeric value to set the x position for displaying
the test result in a plot}

\item{text_y}{A single numeric value to set the y position for displaying
the test result in a plot}
}
\value{
A list of \code{R6} test dataset objects.
}
\description{
The \code{create_usrdata} function creates various types of test datasets.
}
\examples{
## Create a balanced data set with 50 positives and 50 negatives
tset1 <- create_testset("bench", "b100")

## Create an imbalanced data set with 25 positives and 75 negatives
tset2 <- create_testset("bench", "i100")

## Create P1 dataset
tset3 <- create_testset("curve", "c1")

## Create P1 dataset
tset4 <- create_testset("curve", c("c1", "c2"))

}
\seealso{
\code{\link{run_benchmark}} and \code{\link{run_evalcurve}} require
 the list of the datasets generated by this funciton.
 \code{\link{TestDataB}} for benchmarking testdata.
 \code{\link{TestDataC}}, \code{\link{C1DATA}}, \code{\link{C2DATA}},
 and \code{\link{C3DATA}} for curve evaluation testdata.
}

