% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbenchmark.R
\docType{data}
\name{M1DATA}
\alias{M1DATA}
\title{M1: Manually calculated Precision-Recall curves}
\format{A list with 5 items.
\describe{
  \item{scores}{input scores}
  \item{labels}{input labels}
  \item{bp_x}{x values of the calculated curve}
  \item{bp_y}{y values of the calculated curve}
  \item{line_type}{line type between two points.
    \describe{
      \item{"v"}{vertical line}
      \item{"h"}{horizontal line}
      \item{"c"}{non-linear line}
      \item{"p"}{point}
    }
  }
}}
\usage{
data(M1DATA)
}
\description{
A list contains scores and labels of input data, x and y values of
the Precision-Recall curve, and line types between two points.
}
\keyword{datasets}

