% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{data}
\name{M3DATA}
\alias{M3DATA}
\title{M3: Manually calculated Precision-Recall curve}
\format{See \code{\link{M1DATA}}.}
\usage{
data(M3DATA)
}
\description{
A list contains scores and labels of input data, x and y values of
the Precision-Recall curve, and line types between two points.
}
\keyword{datasets}

