% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{data}
\name{P1DATA}
\alias{P1DATA}
\title{P1: Pre-calculated Precision-Recall curve}
\format{A list with 5 items.
\describe{
  \item{scores}{input scores}
  \item{labels}{input labels}
  \item{bp_x}{precalculated recall values for curve evaulation}
  \item{bp_y}{precalculated precision values for curve evaluation}
  \item{tp_x}{x position for displaying the test result in a plot}
  \item{tp_y}{y position for displaying the test result in a plot}
}}
\usage{
data(P1DATA)
}
\description{
A list contains scores, labels, and precalculated recall and precision
values as x and y.
}
\keyword{datasets}

