% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval_results.R
\name{plot_eval_results}
\alias{plot_eval_results}
\title{Plot the result of Precision-Recall curve validation}
\usage{
plot_eval_results(eval_res, base_plot = TRUE, ret_grob = FALSE)
}
\arguments{
\item{base_plot}{A Boolean value to specify whether the base points are
plotted.}

\item{ret_grob}{A Boolean value to specify whether the function returs a
grob object.}

\item{testdata_names}{A character vector to specify the names of test data
sets.}

\item{toolset_name}{A single string to specify the name of tool sets.}
}
\value{
A data frame with validation results.
}
\description{
The \code{plot_eval_results} function validate Precision-Recall curves
   and creat a plot.
}
\examples{
## Plot evaluation results on test datasets r1, r2, and r3
eres1 <- eval_curves(c("r1", "r2", "r3"), "crv")
plot_eval_results(eres1)

}

