% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbenchmark.R
\docType{package}
\name{prcbenchmark}
\alias{prcbenchmark}
\alias{prcbenchmark-package}
\title{prcbenchmark: A package for benchmarking Precision-Recall curves}
\description{
The prcbenchmark package provides three categories of important functions:
tool wrappers, data preparation, and benchmarking.
}
\section{Tool wrappers}{

\code{\link{create_tools}}.
}

\section{Data preparation}{

\code{\link{create_tools}}.
}

\section{Benchmarking}{

\code{\link{create_tools}}.
}

