% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{PRCData}
\alias{PRCData}
\title{Input dataset class for performance evaluation tools}
\format{An R6 class object.}
\usage{
PRCData
}
\description{
\code{PRCData} is a class that contans scores and labels as input data for
  performance evaluation tools.
}
\section{Methods}{

\itemize{
 \item \code{get_datname()}: Get a string of the dataset name.
 \item \code{get_scores()}: Get a vector of predicted scores.
 \item \code{get_labels()}: Get a vector of observed labels.
 \item \code{get_fg()}: Get a vector of positive scores.
 \item \code{get_bg()}: Get a vector of negative scores.
 \item \code{get_fname()}: Get a file name that contains scores and labels.
 \item \code{del_file()}: Get a file name that contains scores and labels.
}
}
\examples{
## An object needs to be instantiated from the class
## before calling any methods
prcdata <- PRCData$new(c(0.1, 0.2, 0.3), c(0, 1, 1), "m1")

}
\keyword{datasets}

