% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prcdata_generator}
\alias{prcdata_generator}
\title{Generate an input dataset class for performance evaluation tools}
\usage{
prcdata_generator()
}
\value{
PRCData \code{R6} class.
}
\description{
The \code{prcdata_generator} function creates an \code{R6} class that
 can be used as an input dataset for performance evaluation tools.
}
\examples{
## Generate input data object
prcdata_cls <- prcdata_generator()

## An object needs to be instantiated from the class
## before calling any methods
prcdata <- prcdata_cls$new(c(0.1, 0.2, 0.3), c(0, 1, 1), "m1")

}
\seealso{
\code{\link{PRCData}} is generated by this function.
}

