% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_interface.R
\name{tool_generator}
\alias{tool_generator}
\title{Wrapper class generator for performace evaluatoin tools}
\usage{
tool_generator(name)
}
\arguments{
\item{name}{A string to specify the name of a performance evaluatoin tool.
The funciton currently accepts the following five tools.
\itemize{
  \item \href{https://rocr.bioinf.mpi-sb.mpg.de/}{ROCR}
  \item \href{http://mark.goadrich.com/programs/AUC/}{AUCCalculator}
  \item \href{https://cran.r-project.org/web/packages/PerfMeas/}{PerfMeas}
  \item \href{https://cran.r-project.org/web/packages/PRROC/}{PRROC}
  \item \href{https://cran.r-project.org/web/packages/precrec/}{precrec}
}}
}
\value{
\code{R6Class} as a wrapper class of performace evaluatoin tools.

\tabular{llll}{
  \strong{Tool name}
  \tab \tab \strong{Generated R6Class} \cr

  ROCR \tab \tab \code{\link{ToolROCR}}   \cr
  AUCCalculator \tab \tab \code{\link{ToolAUCCalculator}} \cr
  PerfMeas \tab \tab \code{\link{ToolPerfMeas}}   \cr
  PRROC \tab \tab \code{\link{ToolPRROC}}   \cr
  precrec \tab \tab \code{\link{Toolprecrec}}
}
}
\description{
The \code{tool_generator} function takes a name of a performance evaluation
  tool and makes an wrapper class for the tool. All generated classes share
  a common interface interited from the \code{\link{ToolBase}} class.
}
\examples{
## Generate a warpper class for ROCR
toolroc_cls <- tool_generator("ROCR")

## An object needs to be instantiated from the class
## before calling any wrapper functions
toolroc <- toolroc_cls$new()

}
\seealso{
\code{\link{ToolBase}} is used as the base class.
  \code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
  \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}},
  and \code{\link{Toolprecrec}} can be genereted by this function.
}

