% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{ToolBase}
\alias{ToolBase}
\title{Base class for the wrapper classes of performace evaluatoin tools}
\format{An R6 class object.}
\usage{
ToolBase
}
\description{
\code{ToolBase} is an abstruct class for the wrapper classes of performace
evaluatoin tools. It is an R6 class that contains nine public methods
for subclass inheritance.
}
\section{Methods}{

\itemize{
 \item \code{call(sdat, retval = TRUE, auc = TRUE)}: It calls an actuall tool
     to calculate Precision-Recall curves.
     \describe{
       \item{\code{sdat}}{\code{R6} object generated by the
                          \code{prepare_data} function.}
       \item{\code{retval}}{A Boolean value to specify whether the method
                            returns the AUC score and x-y positions.}
       \item{\code{auc}}{A Boolean value to specify  whether the AUC score
                        should be calculated.}
      }
 \item \code{get_toolname()}: Get the name of the tool.
 \item \code{get_result()}: Get a list with an AUC score and x-y values.
 \item \code{get_x()}: Get a list with Recall values of a Precision-Recall
                       curve.
 \item \code{get_y()}: Get a list with Precision values of a Precision-Recall
                       curve.
 \item \code{get_auc()}: Get an AUC score.
}
}
\seealso{
\code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
  \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}},
  and \code{\link{Toolprecrec}} are derived from this class.
}
\keyword{datasets}

