% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{create_testdata}
\alias{create_testdata}
\title{Create a list of test datasets}
\usage{
create_testdata(datatype, sampnames = NULL, scores = NULL, labels = NULL,
  dsname = NA, base_x = NULL, base_y = NULL, text_x = NULL,
  text_y = NULL)
}
\arguments{
\item{datatype}{A single string to specify the type of dataset generated by
  this function.

  \describe{
    \item{"random"}{Create test datasets for benchmarking}
    \item{"precalc"}{Create test datasets for curve evaluation}
    \item{"single"}{Create a single test dataset}
  }}

\item{sampnames}{A character vector to specify the names of predefined test
  datasets.

\enumerate{

  \item For benchmarking (\code{datatype = "random"})

  This function uses a name convension for randomly generated data. The
  format is a prefix ('i' or 'b') followed by the number of dataset. The
  prefix 'i' indicates a balanced dataset, whereas 'b' indicates an
  imbalanced dataset. The number can be used with a suffix 'k' or 'm',
  indicating respectively 1000 or 1 million.

  Below are some examples.
  \describe{
    \item{"b100"}{A balanced data set with 50 positives and 50
        negatives.}
    \item{"b10k"}{A balanced data set with 5000 positives and 5000
        negatives.}
    \item{"b1m"}{A balanced data set with 500,000 positives and 500,000
        negatives.}
    \item{"i100"}{An imbalanced data set with 25 positives and 75
        negatives.}
  }

  The function returns a list of \code{\link{TestDataPB}} objects.

  \item For curve evaluation (\code{mode = "precalc"})

  The following three predifned datasets can be specified.

  \tabular{lll}{
    \strong{samp_name}
    \tab \strong{\code{S3} object}
    \tab \strong{data source} \cr

    p1 or P1 \tab \code{\link{TestDataEC}} \tab \code{\link{P1DATA}}   \cr
    p2 or P2 \tab \code{\link{TestDataEC}} \tab \code{\link{P2DATA}}   \cr
    p3 or P3 \tab \code{\link{TestDataEC}} \tab \code{\link{P3DATA}}
  }

  The function returns a list of \code{\link{TestDataEC}} objects.
}}

\item{scores}{A numeric vector to set scores.}

\item{labels}{A numeric vector to set labels.}

\item{dsname}{A single string to specify the name of the dataset.}

\item{base_x}{A numeric vector to set precalculated recall values for
curve evaulation.}

\item{base_y}{A numeric vector to set precalculated precision values for
curve evaulation.}

\item{text_x}{A single numeric value to set the x position for displaying
the test result in a plot}

\item{text_y}{A single numeric value to set the y position for displaying
the test result in a plot}
}
\value{
A list of \code{R6} test dataset objects.
}
\description{
The \code{create_testdata} function creates various types of test datasets.
}
\examples{
## Create a balanced data set with 50 positives and 50 negatives
samp1 <- create_testdata("random", "b100")

## Create an imbalanced data set with 25 positives and 75 negatives
samp2 <- create_testdata("random", "i100")

## Create P1 dataset
samp3 <- create_testdata("precalc", "p1")

## Create P1 dataset
samp4 <- create_testdata("precalc", c("p1", "p2"))

}
\seealso{
\code{\link{run_benchmark}} and \code{\link{eval_curves}} require
 the list of the datasets generated by this funciton.
 \code{\link{TestDataPB}} for benchmarking testdata.
 \code{\link{TestDataEC}}, \code{\link{P1DATA}}, \code{\link{P2DATA}},
 and \code{\link{P3DATA}} for curve evaluation testdata.
}

