% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_interface.R
\name{create_tools}
\alias{create_tools}
\title{Create tool wrapper objects}
\usage{
create_tools(tool_names = c("ROCR", "AUCCalculator", "PerfMeas", "PRROC",
  "precrec"))
}
\arguments{
\item{tool_names}{A character vector to specify the names of
  performance evaluatoin tools. The names for the following five tools can be
  currently used.

  \itemize{
    \item ROCR
    \item AUCCalculator
    \item PerfMeas
    \item PRROC
    \item precrec
  }}
}
\value{
A list of \code{R6} objects.
}
\description{
The \code{create_tools} function takes names of performance evaluation
  tool and generate wrapper objects.
}
\examples{
## Generate all wraper objects
tools1 <- create_tools()

## Generate PRROC and precrec wraper objects
tools2 <- create_tools(c("PRROC", "precrec"))

}
\seealso{
\code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
  \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}},
  and \code{\link{Toolprecrec}} can be genereted by this function.
}

