% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_interface.R
\name{create_toolset}
\alias{create_toolset}
\title{Create a set of tools}
\usage{
create_toolset(set_name)
}
\arguments{
\item{set_name}{A single string to specify the name of a tool set. Following
  three sets are currently available.

  \describe{
    \item{"set1"}{A set optimized for curve calculations}
    \item{"set2"}{A set optimized for AUC calculations}
    \item{"set3"}{A set optimized for both curve and AUC calculations.}
  }}
}
\value{
A list of functions that calls tool calculations.
}
\description{
The \code{create_toolset} function takes a name of tool set and genarates
  a list of wrapper functions for tool calculations.
}
\examples{
## Create a tool set for curve calculations
tool_funcs1 <- create_tools("set1")

## Create a tool set for AUC calculations
tool_funcs2 <- create_tools("set2")

## Create a tool set for both curve and AUC calculations
tool_funcs3 <- create_tools("set3")

}
\seealso{
\code{\link{create_tools}} for creating tool objects.
  \code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
  \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}},
  and \code{\link{Toolprecrec}} as classes of tool objects.
}

