% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_interface.R
\name{create_toolset}
\alias{create_toolset}
\title{Create a set of tools}
\usage{
create_toolset(tool_names = NULL, set_names = NULL, calc_auc = TRUE,
  store_res = TRUE)
}
\arguments{
\item{tool_names}{A character vector to specify the names of
  performance evaluatoin tools. The names for the following five tools can be
  currently used.

  \itemize{
    \item ROCR
    \item AUCCalculator
    \item PerfMeas
    \item PRROC
    \item precrec
  }}

\item{set_names}{A character vector to specify a pre-defined set nam.
  Following six sets are currently available.

  \describe{
    \item{"crv5"}{A set of 5 tools optimized for curve calculations}
    \item{"auc5"}{A set of 5 tools optimized for AUC calculations}
    \item{"both5"}{A set of 5 tools optimized for both curve and AUC
                  calculations.}
    \item{"crv4"}{A set of 4 tools optimized for curve calculations}
    \item{"auc4"}{A set of 4 tools optimized for AUC calculations}
    \item{"both4"}{A set of 4 tools optimized for both curve and AUC
                  calculations.}
  }}

\item{calc_auc}{A Boolean value to specify whether the AUC score should be
calculated.}

\item{store_res}{A Boolean value to specify whether the caluladted curve is
retrieved and stored}
}
\value{
A list of \code{R6} tool objects.
}
\description{
The \code{create_toolset} function takes names of napredefined tools and
  generates a list of wrapper functions for Precision-Recall curve
  calculations.
}
\examples{
## Create ROCR and precrec
tools1 <- create_toolset(c("ROCR", "precrec"))

## Create auc5 tools
tools2 <- create_toolset(set_names = "auc5")

}
\seealso{
\code{\link{run_benchmark}} and \code{\link{run_evalcurve}} require
 the list of the tools generated by this funciton.
 \code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
 \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}}, and
 \code{\link{Toolprecrec}} as R6 tool classes.
}

