% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_evalcurves.R
\name{eval_curves}
\alias{eval_curves}
\title{Evaluate Precision-Recall curves with specified tool sets and test datasets}
\usage{
eval_curves(testdata, tools)
}
\arguments{
\item{testdata}{A character vector to specify a test set generated by
\code{\link{create_testdata}}.}

\item{tools}{A character vector to specify a tool set generated by
\code{\link{create_tools}}.}
}
\value{
A data frame with validation results.
}
\description{
The \code{eval_curves} function run sevearl tests to validate
   Precision-Recall curves.
}
\examples{
## Check Precision-Recall cuvers of a tool set "crv5" on a test
## dataset "r1"
tdat <- create_testdata("precalc", c("p1", "p2", "p3"))
tools <- create_tools(set_names = "crv5")
res1 <- eval_curves(tdat, tools)

}
\seealso{
\code{\link{create_sample}} to generate a test dataset.
   \code{\link{create_tools}} to generate a tool set.
}

