% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_curves.R
\name{eval_curves}
\alias{eval_curves}
\title{Run microbenchmark with a specified set of tools and a sample dataset}
\usage{
eval_curves(testdat_name = "r1", toolset_name = "crv")
}
\arguments{
\item{sdat}{A sample dataset genereated by \code{\link{create_sampleset}} or
\code{\link{create_rnd_sample}}.}

\item{tool_funcs}{A list of functions generated by
\code{\link{create_toolset}}.}

\item{times}{The number of iteration used in
\code{\link[microbenchmark]{microbenchmark}}.}

\item{unit}{A single string to specify the unit used in
\code{\link[microbenchmark]{summary.microbenchmark}}.}
}
\value{
A data frame with microbenchmark results.
}
\description{
The \code{run_benchmark} function runs
    \code{\link[microbenchmark]{microbenchmark}} for a specified set of tools
    and a sample dataset.
}
\examples{
## Check Precision-Recall cuvers of a tool set "crv" on a test dataset "r1"
res1 <- eval_curves("r1", "crv")

}
\seealso{
\code{\link{create_sampleset}} and \code{\link{create_rnd_sample}}
   to generate a sample dataset. \code{\link{create_toolset}} to create a
   set of tools. \code{\link[microbenchmark]{microbenchmark}} for
   benchmarking details.
}

