% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{data}
\name{P2DATA}
\alias{P2DATA}
\title{P2: Manually calculated Precision-Recall curve}
\format{See \code{\link{P1DATA}}.}
\usage{
data(P2DATA)
}
\description{
A list contains scores and labels of input data, x and y values of
the Precision-Recall curve, and line types between two points.
}
\keyword{datasets}

