% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{data}
\name{P3DATA}
\alias{P3DATA}
\title{P2: Pre-calculated Precision-Recall curve}
\format{See \code{\link{P1DATA}}.}
\usage{
data(P3DATA)
}
\description{
A list contains scores, labels, and precalculated recall and precision
values as x and y.
}
\keyword{datasets}

