% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{package}
\name{prcbench}
\alias{prcbench-package}
\alias{prcbench}
\title{prcbench: A package to provide a testing workbench for
  precision-recall curves}
\description{
The prcbench package provides four categories of important functions:
tool interface, test data interface, benchmarking, and curve evaluation.
}
\section{Tool interface}{

The \code{\link{create_toolset}} function creates a common interface for
  seven different tools that calculate Precision-Recall curves. These tools
  are \href{https://ipa-tys.github.io/ROCR/}{ROCR},
  \href{http://mark.goadrich.com/programs/AUC/}{AUCCalculator},
  \href{https://cran.r-project.org/package=PerfMeas}{PerfMeas},
  \href{https://cran.r-project.org/package=PRROC}{PRROC},
  \href{https://cran.r-project.org/package=precrec}{precrec},
  \href{https://yardstick.tidymodels.org/}{yardstick}, and
  \href{https://scikit-learn.org/stable/modules/generated/sklearn.metrics.precision_recall_curve.html}{scikit-learn}.

The \code{sklearn} tool is calculated by a standalone Python module that is
  bundled with \code{prcbench} and derived from the scikit-learn source
  code. It requires \code{reticulate}, a working Python installation and
  \code{numpy}. Without them it returns a flat dummy curve rather than
  raising an error, so the predefined tool sets that contain it stay
  usable.

The \code{\link{create_usrtool}} function helps users to make the same
 interface of the predefined ones for their own tools.
}

\section{Test data interface}{

The \code{\link{create_testset}} function creates two different types of test
  data sets. The first type is for benchmarking, and the second type is for
  curve evaluation.

The \code{\link{create_usrdata}} function helps users to make their own test
  data sets.
}

\section{Benchmarking}{

The \code{\link{run_benchmark}} function takes a tool set and a test data set
  and run \code{\link[microbenchmark]{microbenchmark}} for them.

The timing of the \code{sklearn} tool includes the cost of crossing the
  R/Python boundary, so it is not comparable with the timings of the tools
  written in R.
}

\section{Curve evaluation}{

The \code{\link{run_evalcurve}} function takes a tool set and a test data set
  and evaluates the accuracy of Precision-Recall curves for them.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://evalclass.github.io/prcbench/}
  \item \url{https://github.com/evalclass/prcbench}
  \item Report bugs at \url{https://github.com/evalclass/prcbench/issues}
}

}
\author{
\strong{Maintainer}: Takaya Saito \email{takaya.saito@outlook.com} (\href{https://orcid.org/0000-0002-0154-8452}{ORCID})

Authors:
\itemize{
  \item Takaya Saito \email{takaya.saito@outlook.com} (\href{https://orcid.org/0000-0002-0154-8452}{ORCID})
  \item Marc Rehmsmeier \email{marc.rehmsmeier@ii.uib.no} (\href{https://orcid.org/0000-0002-5021-7721}{ORCID})
}

Other contributors:
\itemize{
  \item The scikit-learn developers (Python code in inst/python, derived from scikit-learn (BSD-3-Clause); see inst/COPYRIGHTS) [copyright holder]
}

}
