% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcbench.R
\docType{package}
\name{prcbench}
\alias{prcbench}
\alias{prcbench-package}
\title{prcbench: A package to provide a testing workbench for Precision-Recall curves}
\description{
The prcbench package provides four categories of important functions:
tool wrappers, data preparation, curve evaluation, and benchmarking.
}
\section{Tool wrappers}{

\code{\link{create_tools}}.
}

\section{Data preparation}{

\code{\link{create_tools}}.
}

\section{Curve evaluation}{

\code{\link{create_tools}}
}

\section{Benchmarking}{

\code{\link{create_tools}}.
}

