% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_benchmark.R
\name{run_benchmark}
\alias{run_benchmark}
\title{Run microbenchmark with specified sets of tools and sets of test datasets}
\usage{
run_benchmark(testdata, tools, times = 5, unit = "ms")
}
\arguments{
\item{testdata}{A character vector to specify a test set generated by
\code{\link{create_testdata}}.}

\item{tools}{A character vector to specify a tool set generated by
\code{\link{create_tools}}.}

\item{times}{The number of iteration used in
\code{\link[microbenchmark]{microbenchmark}}.}

\item{unit}{A single string to specify the unit used in
\code{\link[microbenchmark]{summary.microbenchmark}}.}
}
\value{
A data frame with microbenchmark results.
}
\description{
The \code{run_benchmark} function runs
    \code{\link[microbenchmark]{microbenchmark}} for specified sets of tools
    and sets of test datasets
}
\examples{
## Generate a sample dataset
tdat <- create_testdata("random", c("b10", "i10"))
tools <- create_tools(set_names = c("crv5", "auc5", "both5"))
res1 <- run_benchmark(tdat, tools)

}
\seealso{
\code{\link[microbenchmark]{microbenchmark}} for benchmarking
   details.
   \code{\link{create_sample}} to generate a test dataset.
   \code{\link{create_tools}} to generate a tool set.
}

