% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{run_benchmark}
\alias{run_benchmark}
\title{Run microbenchmark with a specified set of tools and a sample dataset}
\usage{
run_benchmark(sdat, tool_funcs, times = 5, unit = "ms")
}
\arguments{
\item{sdat}{A sample dataset genereated by \code{\link{create_sampleset}} or
\code{\link{create_rnd_sample}}.}

\item{tool_funcs}{A list of functions generated by
\code{\link{create_toolset}}.}

\item{times}{The number of iteration used in
\code{\link[microbenchmark]{microbenchmark}}.}

\item{unit}{A single string to specify the unit used in
\code{\link[microbenchmark]{summary.microbenchmark}}.}
}
\value{
A data frame with microbenchmark results.
}
\description{
The \code{run_benchmark} function runs
    \code{\link[microbenchmark]{microbenchmark}} for a specified set of tools
    and a sample dataset.
}
\examples{
## Generate a sample dataset
samp1 <- create_rnd_sample()
toolset1 <- create_toolset("crv")
res1 <- run_benchmark(samp1, toolset1)

}
\seealso{
\code{\link{create_sampleset}} and \code{\link{create_rnd_sample}}
   to generate a sample dataset. \code{\link{create_toolset}} to create a
   set of tools. \code{\link[microbenchmark]{microbenchmark}} for
   benchmarking details.
}

