% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_benchmark.R
\name{run_benchmark}
\alias{run_benchmark}
\title{Run microbenchmark with specified sets of tools and sets of sample datasets}
\usage{
run_benchmark(samp_names = c("b100", "ib100", "b10k", "ib10k"),
  tool_names = c("crv5", "auc5", "both5"), times = 5, unit = "ms")
}
\arguments{
\item{samp_names}{A character vector to specify the names of sample sets
for \code{\link{create_sample}}.}

\item{tool_names}{A character vector to specify the names of tool sets for
\code{\link{create_tools}}.}

\item{times}{The number of iteration used in
\code{\link[microbenchmark]{microbenchmark}}.}

\item{unit}{A single string to specify the unit used in
\code{\link[microbenchmark]{summary.microbenchmark}}.}
}
\value{
A data frame with microbenchmark results.
}
\description{
The \code{run_benchmark} function runs
    \code{\link[microbenchmark]{microbenchmark}} for specified sets of tools
    and sets of sample datasets
}
\examples{
## Generate a sample dataset
res1 <- run_benchmark(c("b100", "ib100"), c("crv5", "auc5", "both5"))

}
\seealso{
\code{\link[microbenchmark]{microbenchmark}} for benchmarking details.
   \code{\link{create_sample}} to generate sample datasets.
   \code{\link{create_tools}} to create sets of tools.
}

