% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_evalcurves.R
\name{run_evalcurve}
\alias{run_evalcurve}
\title{Evaluate Precision-Recall curves with specified tool sets and test datasets}
\usage{
run_evalcurve(testdata, tools)
}
\arguments{
\item{testdata}{A character vector to specify a test set generated by
\code{\link{create_testset}}.}

\item{tools}{A character vector to specify a tool set generated by
\code{\link{create_toolset}}.}
}
\value{
A data frame with validation results.
}
\description{
The \code{run_evalcurve} function run sevearl tests to validate
   Precision-Recall curves.
}
\examples{
## Check Precision-Recall cuvers of a tool set "crv5" on a test
## dataset "r1"
tdat <- create_testset("precalc", c("p1", "p2", "p3"))
tools <- create_toolset(set_names = "crv5")
res1 <- run_evalcurve(tdat, tools)

}
\seealso{
\code{\link{create_sample}} to generate a test dataset.
   \code{\link{create_toolset}} to generate a tool set.
}

