% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_zzz.R
\docType{class}
\name{TestDataEC}
\alias{TestDataEC}
\title{R6 class of test dataset for Precision-Recall curve evaluation}
\format{An R6 class object.}
\usage{
TestDataEC
}
\description{
\code{TestDataEC} is a class that contans scores and label for performance
  evaluation tools. It provides necessary methods for curve evaluation.
}
\section{Methods}{

\itemize{
 \item \code{set_basepoints_x(x)}: Set precalculated recall values for
   curve evaulation
 \item \code{set_basepoints_y(y)}: Set precalculated precision values for
   curve evaulation
 \item \code{get_basepoints_x()}: Get precalculated recall values for
   curve evaulation
 \item \code{get_basepoints_y()}: Get precalculated precision values for
   curve evaulation
 \item \code{set_textpos_x(x)}: Set the x position for displaying the test
   result in a plot
 \item \code{set_textpos_y(y)}: Set the y position for displaying the test
   result in a plot
 \item \code{get_textpos_x()}: Get the x position for displaying the test
   result in a plot
 \item \code{get_textpos_y()}: Get the y position for displaying the test
   result in a plot
}

Following seven methods are interited from \code{\link{TestDataPB}}. See
\code{\link{TestDataPB}} for the method descriptions.
\itemize{
  \item \code{get_datname()}
  \item \code{get_scores()}
  \item \code{get_labels()}
  \item \code{get_fg()}
  \item \code{get_bg()}
  \item \code{get_fname()}
  \item \code{del_file()}
}
}
\examples{
## Initialize with socres, labels, and a dataset name
testdata <- TestDataEC$new(c(0.1, 0.2, 0.3), c(0, 1, 1), "m1")

## create_testset should be used for benchmarking
testdata2 <- create_testset("single", scores = c(0.1, 0.2, 0.3),
                            labels = c(0, 1, 1), dsname = "m1")

}
\seealso{
\code{\link{create_testset}} for creating a list of test datasets.
  It is derived from \code{\link{TestDataPB}}.
}
\keyword{datasets}

