% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{ToolAUCCalculator}
\alias{ToolAUCCalculator}
\title{R6 class of the AUCCalculator tool}
\format{An R6 class object.}
\usage{
ToolAUCCalculator
}
\description{
\code{ToolAUCCalculator} is a wrapper class for
the \href{http://mark.goadrich.com/programs/AUC/}{AUCCalculator} tool, which
is a Java library that provides calculatoins of ROC and Precision-Recall
curves.
}
\section{Inheritance}{

\code{\link{ToolIFBase}}
}

\section{Methods}{


\describe{
  \item{\code{set_java_call(type, fpath = NULL)}}{
    It sets an AUCCalculator jar file.

    \describe{
      \item{\code{type}}{A string to specify the wrapper method. It must be
                         either "syscall" or "rjava". It uses a system call
                         and \code{\link[rJava]{.jcall}} of the \code{rJava}
                         library, respectively.}
      \item{\code{fpath}}{File path of the AUCCalculator jar file,
                          e.g. \code{"/path1/path2/auc.jar"}.}
    }
  }
}

Following seven methods are interited from \code{\link{ToolIFBase}}. See
\code{\link{ToolIFBase}} for the method descriptions.
\itemize{
  \item \code{call((testdata, calc_auc, store_res)}
  \item \code{get_toolname()}
  \item \code{get_setname()}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## Initialization
toolauccalc <- ToolAUCCalculator$new()

## Show object info
toolauccalc

## create_toolset should be used for benchmarking and curve evalation
toolauccalc2 <- create_toolset("AUCCalculator")

}
\seealso{
This class is derived from \code{\link{ToolIFBase}}.
   \code{\link{create_toolset}} for creating a list of tools.
}
\keyword{datasets}

