% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{ToolBase}
\alias{ToolBase}
\title{Base class for the wrapper classes of performace evaluatoin tools}
\format{\code{R6Class} (abstruct class)}
\usage{
It must be subclassed.
}
\description{
ToolBase is an abstruct class for the wrapper classes of performace
evaluatoin tools. It is an R6Class that contains nine public methods
for subclass inheritance.
}
\section{Methods}{

\describe{
  \item{\code{call(sdat, retval = TRUE, auc = TRUE)}}{
    It calls an actuall tool to calculate Precision-Recall curves.

    \describe{
      \item{\code{sdat}}{\code{R6Class} generated by the \code{prepare_data}
                         function.}
      \item{\code{retval}}{A Boolean value to specify whether the method
                           returns the AUC score and x-y positions.}
      \item{\code{auc}}{A Boolean value to specify  whether the AUC score
                        should be calculated.}
     }
   }

  \item{\code{set_result(val)}}{
    Store an AUC score and x-y values specified with \code{val}. The format
    of \code{val} must be \code{list(x = list(), y = list(),
    auc = numeric())}.
  }

  \item{\code{get_result()}}{
    Get a list with an AUC score and x-y values.
  }

  \item{\code{set_x(val)}}{
    Store x values specified with \code{val}. \code{val} must be a list with
    Recall values of a Precision-Recall curve.
  }

  \item{\code{get_x()}}{
    Get a list with Recall values of a Precision-Recall curve.
  }

  \item{\code{set_y(val)}}{
    Store y values specified with \code{val}. \code{val} must be a list with
    Precision values of a Precision-Recall curve.
  }

  \item{\code{get_y()}}{
    Get a list with Precision values of a Precision-Recall curve.
  }

  \item{\code{set_auc(val)}}{
    Store an AUC score
  }

  \item{\code{get_auc()}}{
     Get an AUC score
  }
}
}
\seealso{
\code{\link{ToolROCR}}, \code{\link{ToolAUCCalculator}},
  \code{\link{ToolPerfMeas}}, \code{\link{ToolPRROC}},
  and \code{\link{Toolprecrec}} are derived from this class.
  \code{\link{tool_generator}} is a generator function that uses this class.
}
\keyword{data}

