% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{ToolPerfMeas}
\alias{ToolPerfMeas}
\title{PerfMeas wrapper class}
\format{\code{R6}}
\usage{
ToolPerfMeas$new()
}
\description{
\code{ToolPerfMeas} is a wrapper class for
\href{https://cran.r-project.org/web/packages/PerfMeas/}{PerfMeas},
which is an R library that provides several performance measures. It is
generated by a generator function as \code{tool_generator("PerfMeas")}.
}
\section{Inheritance}{

\code{\link{ToolBase}}
}

\section{Methods}{

Following five methods are interited from \code{\link{ToolBase}}. See
\code{\link{ToolBase}} for the method descriptions.
\itemize{
  \item \code{call(sdat, retval = TRUE, auc = TRUE)}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## An object needs to be instantiated from the class
## before calling any wrapper functions
toolpm1 <- ToolPerfMeas$new()

## tool_generator can also be used to create a wrapper object
toolpm2 <- tool_generator("PerfMeas")$new()

}
\keyword{data}

