% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{ToolPerfMeas}
\alias{ToolPerfMeas}
\title{R6 class of the PerfMeas tool}
\format{An R6 class object.}
\usage{
ToolPerfMeas
}
\description{
\code{ToolPerfMeas} is a wrapper class for
the \href{https://cran.r-project.org/web/packages/PerfMeas/}{PerfMeas} tool,
which is an R library that provides several performance measures.
}
\section{Inheritance}{

\code{\link{ToolIFBase}}
}

\section{Methods}{

Following seven methods are interited from \code{\link{ToolIFBase}}. See
\code{\link{ToolIFBase}} for the method descriptions.
\itemize{
  \item \code{call(testdata, calc_auc, store_res)}
  \item \code{get_toolname()}
  \item \code{get_setname()}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## Initialization
toolperf <- ToolPerfMeas$new()

## Show object info
toolperf

## create_tools should be used for benchmarking and curve evalation
toolperf2 <- create_tools("PerfMeas")

}
\seealso{
This class is derived from \code{\link{ToolIFBase}}.
   \code{\link{create_tools}} for creating a list of tools.
}
\keyword{datasets}

