% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{ToolPerfMeas}
\alias{ToolPerfMeas}
\title{PerfMeas wrapper class}
\format{An R6 class object.}
\usage{
ToolPerfMeas
}
\description{
\code{ToolPerfMeas} is a wrapper class for
\href{https://cran.r-project.org/web/packages/PerfMeas/}{PerfMeas},
which is an R library that provides several performance measures.
}
\section{Inheritance}{

\code{\link{ToolBase}}
}

\section{Methods}{

Following five methods are interited from \code{\link{ToolBase}}. See
\code{\link{ToolBase}} for the method descriptions.
\itemize{
  \item \code{call(sdat, retval = TRUE, auc = TRUE)}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## An object needs to be instantiated from the class
## before calling any wrapper functions
toolpm1 <- ToolPerfMeas$new()

}
\seealso{
This class is derived from \code{\link{ToolBase}}. The object
can be also instantiated by \code{\link{create_tools}("PerfMeas")}.
}
\keyword{datasets}

