% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{ToolPerfMeas}
\alias{ToolPerfMeas}
\title{R6 class of the PerfMeas tool}
\format{An R6 class object.}
\usage{
ToolPerfMeas
}
\description{
\code{ToolPerfMeas} is a wrapper class for
the \href{https://cran.r-project.org/package=PerfMeas}{PerfMeas} tool,
which is an R library that provides several performance measures.
}
\section{Inheritance}{

\code{\link{ToolIFBase}}
}

\section{Methods}{

Following nine methods are inherited from \code{\link{ToolIFBase}}. See
\code{\link{ToolIFBase}} for the method descriptions.
\itemize{
  \item \code{call(testset, calc_auc, store_res)}
  \item \code{get_toolname()}
  \item \code{set_toolname(toolname)}
  \item \code{get_setname()}
  \item \code{set_setname(setname)}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## Initialization
toolperf <- ToolPerfMeas$new()

## Show object info
toolperf

## create_toolset should be used for benchmarking and curve evaluation
toolperf2 <- create_toolset("PerfMeas")

}
\seealso{
This class is derived from \code{\link{ToolIFBase}}.
   \code{\link{create_toolset}} for creating a list of tools.
}
\keyword{datasets}

