% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{Toolprecrec}
\alias{Toolprecrec}
\title{precrec wrapper class}
\format{\code{R6}}
\usage{
Toolprecrec$new()
}
\description{
\code{Toolprecrec} is a wrapper class for
\href{https://cran.r-project.org/web/packages/precrec/}{precrec}, which is an
R library that provides calculatoins of ROC and Precision-Recall curves. It
is generated by a generator function as \code{tool_generator("precrec")}.
}
\section{Inheritance}{

\code{\link{ToolBase}}
}

\section{Methods}{

Following five methods are interited from \code{\link{ToolBase}}. See
\code{\link{ToolBase}} for the method descriptions.
\itemize{
  \item \code{call(sdat, retval = TRUE, auc = TRUE)}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## An object needs to be instantiated from the class
## before calling any wrapper functions
toolprecrec1 <- Toolprecrec$new()

## tool_generator can also be used to create a wrapper object
toolprecrec2 <- tool_generator("precrec")$new()

}
\keyword{data}

