% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{Toolprecrec}
\alias{Toolprecrec}
\title{precrec wrapper class}
\format{An R6 class object.}
\usage{
Toolprecrec
}
\description{
\code{Toolprecrec} is a wrapper class for
\href{https://cran.r-project.org/web/packages/precrec/}{precrec}, which is an
R library that provides calculatoins of ROC and Precision-Recall curves. It
is generated by a generator function as \code{tool_generator("precrec")}.
}
\section{Inheritance}{

\code{\link{ToolBase}}
}

\section{Methods}{

Following six methods are interited from \code{\link{ToolBase}}. See
\code{\link{ToolBase}} for the method descriptions.
\itemize{
  \item \code{call(sdat, retval = TRUE, auc = TRUE)}
  \item \code{get_toolname()}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## An object needs to be instantiated from the class
## before calling any wrapper functions
toolprecrec1 <- Toolprecrec$new()

}
\seealso{
This class is derived from \code{\link{ToolBase}}. The object
can be also instantiated by \code{\link{create_tools}("precrec")}.
}
\keyword{datasets}

