% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{Toolprecrec}
\alias{Toolprecrec}
\title{R6 class of the precrec tool}
\format{An R6 class object.}
\usage{
Toolprecrec
}
\description{
\code{Toolprecrec} is a wrapper class for
the \href{https://cran.r-project.org/package=precrec}{precrec} tool,
which is an R library that provides calculations of ROC and Precision-Recall
curves.
}
\section{Inheritance}{

\code{\link{ToolIFBase}}
}

\section{Methods}{

Following nine methods are inherited from \code{\link{ToolIFBase}}. See
\code{\link{ToolIFBase}} for the method descriptions.
\itemize{
  \item \code{call(testset, calc_auc, store_res)}
  \item \code{get_toolname()}
  \item \code{set_toolname(toolname)}
  \item \code{get_setname()}
  \item \code{set_setname(setname)}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## Initialization
toolprecrec <- Toolprecrec$new()

## Show object info
toolprecrec

## create_toolset should be used for benchmarking and curve evaluation
toolprecrec2 <- create_toolset("precrec")

}
\seealso{
This class is derived from \code{\link{ToolIFBase}}.
   \code{\link{create_toolset}} for creating a list of tools.
}
\keyword{datasets}

