% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{ToolPRROC}
\alias{ToolPRROC}
\title{PRROC wrapper class}
\format{\code{R6Class}}
\usage{
ToolPRROC$new()
}
\description{
\code{ToolPRROC} is a wrapper class for
\href{https://cran.r-project.org/web/packages/PRROC/}{PRROC}, which is an R
library that provides calculatoins of ROC and Precision-Recall curves. It
is generated by a generator function as
\code{tool_generator("PRROC")}.
}
\section{Inheritance}{

\code{\link{ToolBase}}
}

\section{Methods}{


\describe{
  \item{\code{set_curve(val)}}{A Boolean value to specifiy whether a
                               Precision-Recall curve is calculated.}
  \item{\code{set_minStepSize(val)}}{A numeric value to specify the minimum
                                     step size between itermediate points.}
}

Following nine methods are interited from \code{\link{ToolBase}}. See
\code{\link{ToolBase}} for the method descriptions.
\itemize{
  \item \code{call(sdat, retval = TRUE, auc = TRUE)}
  \item \code{set_result(val)}
  \item \code{get_result()}
  \item \code{set_x(val)}
  \item \code{get_x()}
  \item \code{set_y(val)}
  \item \code{get_y()}
  \item \code{set_auc(val)}
  \item \code{get_auc()}
}
}
\examples{
## An object needs to be instantiated from the class
## before calling any wrapper functions
toolprroc1 <- ToolPRROC$new()

## tool_generator can also be used to create a wrapper object
toolprroc2 <- tool_generator("PRROC")$new()

}
\keyword{data}

