% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{ToolPRROC}
\alias{ToolPRROC}
\title{PRROC wrapper class}
\format{An R6 class object.}
\usage{
ToolPRROC
}
\description{
\code{ToolPRROC} is a wrapper class for
\href{https://cran.r-project.org/web/packages/PRROC/}{PRROC}, which is an R
library that provides calculatoins of ROC and Precision-Recall curves. It
is generated by a generator function as
\code{tool_generator("PRROC")}.
}
\section{Inheritance}{

\code{\link{ToolBase}}
}

\section{Methods}{

\describe{
  \item{\code{set_curve(val)}}{A Boolean value to specifiy whether a
                               Precision-Recall curve is calculated.}
  \item{\code{set_minStepSize(val)}}{A numeric value to specify the minimum
                                     step size between itermediate points.}
}

Following five methods are interited from \code{\link{ToolBase}}. See
\code{\link{ToolBase}} for the method descriptions.
\itemize{
  \item \code{call(sdat, retval = TRUE, auc = TRUE)}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## An object needs to be instantiated from the class
## before calling any wrapper functions
toolprroc1 <- ToolPRROC$new()

}
\seealso{
This class is derived from \code{\link{ToolBase}}.
}
\keyword{datasets}

