% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{ToolPRROC}
\alias{ToolPRROC}
\title{ToolPRROC}
\format{
An \code{R6} class object.
}
\description{
\code{R6} class of the PRROC tool
}
\details{
\code{ToolPRROC} is a wrapper class for
the \href{https://cran.r-project.org/package=PRROC}{PRROC} tool, which
is an R library that provides calculations of ROC and Precision-Recall
curves.
}
\examples{
## Initialization
toolprroc <- ToolPRROC$new()

## Show object info
toolprroc

## create_toolset should be used for benchmarking and curve evaluation
toolprroc2 <- create_toolset("PRROC")

}
\seealso{
This class is derived from \code{\link{ToolIFBase}}.
   \code{\link{create_toolset}} for creating a list of tools.
}
\section{Super class}{
\code{\link[prcbench:ToolIFBase]{prcbench::ToolIFBase}} -> \code{ToolPRROC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ToolPRROC$new()}}
\item \href{#method-set_curve}{\code{ToolPRROC$set_curve()}}
\item \href{#method-set_minStepSize}{\code{ToolPRROC$set_minStepSize()}}
\item \href{#method-set_aucType}{\code{ToolPRROC$set_aucType()}}
\item \href{#method-clone}{\code{ToolPRROC$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="call">}\href{../../prcbench/html/ToolIFBase.html#method-call}{\code{prcbench::ToolIFBase$call()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_auc">}\href{../../prcbench/html/ToolIFBase.html#method-get_auc}{\code{prcbench::ToolIFBase$get_auc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_result">}\href{../../prcbench/html/ToolIFBase.html#method-get_result}{\code{prcbench::ToolIFBase$get_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_setname">}\href{../../prcbench/html/ToolIFBase.html#method-get_setname}{\code{prcbench::ToolIFBase$get_setname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_toolname">}\href{../../prcbench/html/ToolIFBase.html#method-get_toolname}{\code{prcbench::ToolIFBase$get_toolname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_x">}\href{../../prcbench/html/ToolIFBase.html#method-get_x}{\code{prcbench::ToolIFBase$get_x()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_y">}\href{../../prcbench/html/ToolIFBase.html#method-get_y}{\code{prcbench::ToolIFBase$get_y()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="print">}\href{../../prcbench/html/ToolIFBase.html#method-print}{\code{prcbench::ToolIFBase$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="set_setname">}\href{../../prcbench/html/ToolIFBase.html#method-set_setname}{\code{prcbench::ToolIFBase$set_setname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="set_toolname">}\href{../../prcbench/html/ToolIFBase.html#method-set_toolname}{\code{prcbench::ToolIFBase$set_toolname()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Default class initialization method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToolPRROC$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{set value for \code{curve}, \code{minStepSize}, \code{aucType}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_curve"></a>}}
\if{latex}{\out{\hypertarget{method-set_curve}{}}}
\subsection{Method \code{set_curve()}}{
A Boolean value to specify whether precision-recall curve is calculated.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToolPRROC$set_curve(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{TRUE: calculate, FALSE: not calculate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_minStepSize"></a>}}
\if{latex}{\out{\hypertarget{method-set_minStepSize}{}}}
\subsection{Method \code{set_minStepSize()}}{
A numeric value to specify the minimum step size between two intermediate points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToolPRROC$set_minStepSize(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Step size between two  points.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_aucType"></a>}}
\if{latex}{\out{\hypertarget{method-set_aucType}{}}}
\subsection{Method \code{set_aucType()}}{
Set the AUC calculation method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToolPRROC$set_aucType(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{1: integral, 2: Davis Goadrich}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToolPRROC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
