% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{ToolROCR}
\alias{ToolROCR}
\title{ROCR wrapper class}
\format{\code{R6}}
\usage{
ToolROCR$new()
}
\description{
\code{ToolROCR} is a wrapper class for
\href{https://rocr.bioinf.mpi-sb.mpg.de/}{ROCR}, which is an R library
that provides calculatoins of various performance evaluation measures. It
is generated by a generator function as \code{tool_generator("ROCR")}.
}
\section{Inheritance}{

\code{\link{ToolBase}}
}

\section{Methods}{

Following five methods are interited from \code{\link{ToolBase}}. See
\code{\link{ToolBase}} for the method descriptions.
\itemize{
  \item \code{call(sdat, retval = TRUE, auc = TRUE)}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## An object needs to be instantiated from the class
## before calling any wrapper functions
toolroc1 <- ToolROCR$new()

## tool_generator can also be used to create a wrapper object
toolroc2 <- tool_generator("ROCR")$new()

}
\seealso{
This class is derived from \code{\link{ToolBase}}.
\code{\link{tool_generator}} is a generator function for this class.
}
\keyword{data}

