% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{ToolROCR}
\alias{ToolROCR}
\title{R6 class of the ROCR tool}
\format{An R6 class object.}
\usage{
ToolROCR
}
\description{
\code{ToolROCR} is a wrapper class for
the \href{https://rocr.bioinf.mpi-sb.mpg.de/}{ROCR} tool, which is an R
library that provides calculations of various performance evaluation
measures.
}
\section{Inheritance}{

\code{\link{ToolIFBase}}
}

\section{Methods}{

Following seven methods are inherited from \code{\link{ToolIFBase}}. See
\code{\link{ToolIFBase}} for the method descriptions.
\itemize{
  \item \code{call(testset, calc_auc, store_res)}
  \item \code{get_toolname()}
  \item \code{get_setname()}
  \item \code{get_result()}
  \item \code{get_x()}
  \item \code{get_y()}
  \item \code{get_auc()}
}
}
\examples{
## Initialization
toolrocr <- ToolROCR$new()

## Show object info
toolrocr

## create_toolset should be used for benchmarking and curve evaluation
toolrocr2 <- create_toolset("ROCR")

}
\seealso{
This class is derived from \code{\link{ToolIFBase}}.
   \code{\link{create_toolset}} for creating a list of tools.
}
\keyword{datasets}

