% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_zzz.R
\docType{class}
\name{Toolsklearn}
\alias{Toolsklearn}
\title{Toolsklearn}
\format{
An \code{R6} class object.
}
\description{
\code{R6} class of the scikit-learn tool
}
\details{
\code{Toolsklearn} is a wrapper class for the precision-recall curve
  calculation of
  \href{https://scikit-learn.org/stable/modules/generated/sklearn.metrics.precision_recall_curve.html}{scikit-learn},
  which is a machine learning library for Python.

The calculation is performed by a standalone Python module that is bundled
  with \code{prcbench} and derived from the scikit-learn source code. As a
  result, scikit-learn itself is not required, but \code{reticulate},
  a working Python installation, and \code{numpy} are. The tool can be
  created without them, whereas the actual calculation cannot be performed.
  In that case the tool returns a flat dummy curve instead of raising an
  error, in the same way as \code{\link{ToolAUCCalculator}} does without
  \code{rJava}, so that the predefined tool sets keep working on a machine
  without Python.

Two AUC calculation methods are available. \code{aucType = 1} uses average
  precision, which is the summary scikit-learn recommends for
  precision-recall curves, whereas \code{aucType = 2} uses the trapezoidal
  rule. The scikit-learn documentation discourages the use of the
  trapezoidal rule for precision-recall curves.
}
\examples{
## Initialization
toolsklearn <- Toolsklearn$new()

## Show object info
toolsklearn

## create_toolset should be used for benchmarking and curve evaluation
toolsklearn2 <- create_toolset("sklearn")

}
\seealso{
This class is derived from \code{\link{ToolIFBase}}.
   \code{\link{create_toolset}} for creating a list of tools.
}
\section{Super class}{
\code{\link[prcbench:ToolIFBase]{ToolIFBase}} -> \code{Toolsklearn}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Toolsklearn-initialize}{\code{Toolsklearn$new()}}
    \item \href{#method-Toolsklearn-set_drop_intermediate}{\code{Toolsklearn$set_drop_intermediate()}}
    \item \href{#method-Toolsklearn-set_aucType}{\code{Toolsklearn$set_aucType()}}
    \item \href{#method-Toolsklearn-clone}{\code{Toolsklearn$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="call"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-call'><code>ToolIFBase$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_auc"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-get_auc'><code>ToolIFBase$get_auc()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_result"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-get_result'><code>ToolIFBase$get_result()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_setname"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-get_setname'><code>ToolIFBase$get_setname()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_toolname"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-get_toolname'><code>ToolIFBase$get_toolname()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_x"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-get_x'><code>ToolIFBase$get_x()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="get_y"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-get_y'><code>ToolIFBase$get_y()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="print"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-print'><code>ToolIFBase$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="set_setname"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-set_setname'><code>ToolIFBase$set_setname()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="prcbench" data-topic="ToolIFBase" data-id="set_toolname"><a href='../../prcbench/html/ToolIFBase.html#method-ToolIFBase-set_toolname'><code>ToolIFBase$set_toolname()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toolsklearn-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Toolsklearn-initialize}{}}}
\subsection{\code{Toolsklearn$new()}}{
  Default class initialization method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Toolsklearn$new(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{set value for \code{drop_intermediate}, \code{aucType}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toolsklearn-set_drop_intermediate"></a>}}
\if{latex}{\out{\hypertarget{method-Toolsklearn-set_drop_intermediate}{}}}
\subsection{\code{Toolsklearn$set_drop_intermediate()}}{
  A Boolean value to specify whether suboptimal thresholds are dropped.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Toolsklearn$set_drop_intermediate(val)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{val}}{TRUE: drop, FALSE: keep.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toolsklearn-set_aucType"></a>}}
\if{latex}{\out{\hypertarget{method-Toolsklearn-set_aucType}{}}}
\subsection{\code{Toolsklearn$set_aucType()}}{
  Set the AUC calculation method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Toolsklearn$set_aucType(val)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{val}}{1: average precision, 2: trapezoidal rule}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toolsklearn-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Toolsklearn-clone}{}}}
\subsection{\code{Toolsklearn$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Toolsklearn$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
